% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestKeeper.R
\name{bestKeeper}
\alias{bestKeeper}
\title{Determines stability of genes}
\usage{
bestKeeper(expression, ctVal = TRUE)
}
\arguments{
\item{expression}{a matrix of expression levels. Each row corresponds to a sample and each column to a gene.}

\item{ctVal}{a logical value indicating data type. If ct-values are input, ctVal=TRUE, otherwise, ctVal=FALSE.}
}
\value{
A list containing CP.statistics, pair.Wise.cor and HKG.vs.BestKeeper, which are returned by cpSta(), pearsonCor() and bki(), respectively.
}
\description{
This function combines the results of cpSta(), pearsonCor() and bki().
}
\examples{
FIBct
bestKeeper(FIBct)
}
\references{
Pfaffl MW, Tichopad A, Prgomet C, Neuvians TP. Biotechnol Lett (2004) <doi: 10.1023/B:BILE.0000019559.84305.47>
}
