% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNorm.R
\name{FIB}
\alias{FIB}
\title{Reload Saved Datasets}
\description{
The normalized expression level of the ten housekeeping genes in fibroblast cells
}
\details{
The normalized expression level of the ten housekeeping genes in fibroblast cells
}
\references{
Vandesompele J, De Preter K, Pattyn F, Poppe B, Van Roy N, De Paepe A, Speleman F (2002) <doi: 10.1186/gb-2002-3-7-research0034>.
}
