% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNorm.R
\name{FIBct}
\alias{FIBct}
\title{Reload Saved Datasets}
\description{
The CT values of the ten housekeeping genes in fibroblast cells
}
\details{
The CT values of the ten housekeeping genes in fibroblast cells
}
\references{
Vandesompele J, De Preter K, Pattyn F, Poppe B, Van Roy N, De Paepe A, Speleman F (2002) <doi: 10.1186/gb-2002-3-7-research0034>.
}
