% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNorm.R
\name{pairwiseV}
\alias{pairwiseV}
\title{Calculates V(n+1/n) values}
\usage{
pairwiseV(expression, ctVal = TRUE)
}
\arguments{
\item{expression}{a matrix of expression levels. Each row corresponds to a sample and each column to a gene.}

\item{ctVal}{a logical value indicating data type. If ct-values are input, ctVal=TRUE, otherwise, ctVal=FALSE.}
}
\value{
A Series of values [V2/3, V3/V4, V4/V5, ...].
}
\description{
Useful for establishing the quality of your normalization regime. See Vandesompele 2002 for advice on interpretation.
}
\examples{
FIB
pairwiseV(FIB,ctVal=FALSE)
FIBct
pairwiseV(FIBct)
}
\references{
Vandesompele J, De Preter K, Pattyn F, Poppe B, Van Roy N, De Paepe A, Speleman F (2002) <doi: 10.1186/gb-2002-3-7-research0034>.
}
