% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNorm.R
\name{plotM}
\alias{plotM}
\title{Plots average M of remaining genes}
\usage{
plotM(Mrem)
}
\arguments{
\item{Mrem}{the result returned by function of geNorm()}
}
\description{
This function plots the average expression stability values of remaining control genes.
}
\examples{
FIB
x=geNorm(FIB,ctVal=FALSE)
plotM(x)

FIBct
y=geNorm(FIBct)
plotM(y)
}
\references{
Vandesompele J, De Preter K, Pattyn F, Poppe B, Van Roy N, De Paepe A, Speleman F (2002) <doi: 10.1186/gb-2002-3-7-research0034>.
}
