% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNorm.R
\name{plotV}
\alias{plotV}
\title{Plots V(n+1/n) values}
\usage{
plotV(Vs)
}
\arguments{
\item{Vs}{the result returned by function of pairwiseV()}
}
\description{
This function plots the average expression stability values of remaining control genes.
}
\examples{
FIB
Vs1=pairwiseV(FIB,ctVal=F)
plotV(Vs1)

FIBct
Vs2=pairwiseV(FIBct)
plotV(Vs2)
}
\references{
Vandesompele J, De Preter K, Pattyn F, Poppe B, Van Roy N, De Paepe A, Speleman F (2002) <doi: 10.1186/gb-2002-3-7-research0034>.
}
